Added Frontend code to the project and build Authentiation and Authoriation

+++ Task-1/src/main/java/com/task1/entity/AddressClass.java
@@ -12,7 +12,7 @@ public class AddressClass {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userAddressId;
+    private long userAddressId;
 
     @Column(nullable = false)
     private String userHouseNumber;
